make sync_diff_inspector pkg importable
